lecsum-be/requirements.txt
# Web Framework
fastapi==0.124.4
uvicorn==0.38.0
python-multipart==0.0.20

# LLM / LangChain
openai==2.11.0
langchain==1.1.3
langchain-openai==1.1.3
langchain-community==0.4.1
langchain-core==1.2.0
langchain-chroma==1.1.0
langchain-text-splitters==1.1.0

# Vector DB
chromadb==1.3.6

# File Processing
pypdf==6.4.1
python-pptx==0.6.23

# Validation / Settings
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1

# Database (MySQL)
SQLAlchemy==2.0.45
pymysql==1.1.1

# Dev
watchfiles==1.1.1